refactor livechat fallback
